+++ libs/base_core/src/data/mod.rs
@@ -6,6 +6,7 @@
 #![cfg_attr(base_safe_data, forbid(unsafe_code))]
 
 mod bit; // Bitwise, bitfield!
+mod handle; // define_handle!
 mod sort; // Sort
 
 pub mod codec;
@@ -16,7 +17,7 @@ pub mod list;
 
 crate::structural_mods! { // _mods, _pub_mods, _workspace_internals
     _mods {
-        pub use super::{bit::_all::*, sort::_all::*};
+        pub use super::{bit::_all::*, handle::*, sort::_all::*};
     }
     _pub_mods {
         #[doc(inline)]