+++ libs/base_core/src/sys/mem/arena/define.rs
@@ -6,7 +6,11 @@
 #[cfg(any(doc, test))]
 define_arena! {
     #[doc = crate::_TAG_EXAMPLE!()]
-    pub ExampleArena
+    /// An example memory arena. Generated with [`define_arena!`].
+    pub ExampleArena;
+    /// An example memory arena handle. Generated with [`define_arena!`].
+    #[doc = crate::_TAG_EXAMPLE!()]
+    pub ExampleArenaHandle;
 }
 
 /// A custom memory arena generator.
@@ -23,37 +27,36 @@ macro_rules! define_arena {
     // Point of entry, defines the names
     (
     $(#[$arena_attr:meta])*
-    $vis:vis $Arena:ident
+    $vis:vis $Arena:ident;
+    $(#[$handle_attr:meta])*
+    $hvis:vis $Handle:ident $(;)?
     ) => {
-        define_arena![%arena
+        $crate::define_handle![$(#[$handle_attr])* $hvis $Handle];
+        $crate::define_arena![%arena
             $(#[$arena_attr])* $vis $Arena,
-            $crate::ArenaHandle,
+            $Handle,
             $crate::__Arena::<CAP>
         ];
     };
-    () => {};
     // calls the necessary arms in order.
     (%arena $(#[$arena_attr:meta])* $vis:vis $Arena:ident, $Handle:ty, $_:ty) => { $crate::paste! {
-        define_arena![%main
+        $crate::define_arena![%main
             $(#[$arena_attr])* $vis $Arena, // the arena type
             $Handle, // the handle type
             [<$Arena Mark>], // the mark type
             $_ // the internal ops arena namespace
         ];
-        define_arena![%prim $Arena, $Handle, ($)];
-        define_arena![%mark $Arena, $(#[$arena_attr])* $vis [<$Arena Mark>]];
+        $crate::define_arena![%prim $vis $Arena, $Handle, ($)];
+        $crate::define_arena![%mark $Arena, $(#[$arena_attr])* $vis [<$Arena Mark>]];
 
         #[cfg(test)]
-        define_arena![%tests $Arena, [<test_ $Arena>]];
+        $crate::define_arena![%tests $Arena, [<test_ $Arena>]];
     }};
     // $Arena:   the name of the new generated arena.
     // $Handle:  the handle type used by the arena.
     // $_:       internal arena helper.
     (%main $(#[$arena_attr:meta])* $vis:vis $Arena:ident, $Handle:ty, $Mark:ty, $_:ty) => {
         $(#[$arena_attr])*
-        /// An heterogeneous byte arena.
-        ///
-        /// Generated with [`define_arena!`].
         #[derive(Clone, Debug)]
         $vis struct $Arena<const CAP: usize> {
             data: [$crate::MaybeByte; CAP],
@@ -71,34 +74,34 @@ macro_rules! define_arena {
         }
 
         // Fundamental methods.
-        #[allow(dead_code)]
+        #[allow(dead_code, private_interfaces)]
         impl<const CAP: usize> $Arena<CAP> {
 
             /// Returns a new empty arena.
             #[inline(always)]
-            pub const fn new() -> Self {
+            $vis const fn new() -> Self {
                 Self { data: <$_>::new_array(), len: 0 }
             }
 
             #[inline(always)]
             /// Returns the total capacity.
-            pub const fn capacity(&self) -> usize { CAP }
+            $vis const fn capacity(&self) -> usize { CAP }
             #[inline(always)]
             /// Return the occupied length.
-            pub const fn len(&self) -> usize { self.len }
+            $vis const fn len(&self) -> usize { self.len }
             #[inline(always)]
             /// Whether the arena is empty.
-            pub const fn is_empty(&self) -> bool { self.len == 0 }
+            $vis const fn is_empty(&self) -> bool { self.len == 0 }
             #[inline(always)]
             /// Returns the remaining byte capacity.
-            pub const fn remaining(&self) -> usize { CAP - self.len }
+            $vis const fn remaining(&self) -> usize { CAP - self.len }
             /// Returns `true` if n bytes fit in the remaining capacity.
             #[inline(always)]
-            pub const fn can_write(&self, n: usize) -> bool { self.len + n <= CAP }
+            $vis const fn can_write(&self, n: usize) -> bool { self.len + n <= CAP }
 
             #[inline(always)]
             /// Compares two arenas for equality.
-            pub const fn eq(&self, other: &Self) -> bool {
+            $vis const fn eq(&self, other: &Self) -> bool {
                 $crate::Slice::<u8>::eq(self.as_bytes(), other.as_bytes())
             }
 
@@ -106,26 +109,26 @@ macro_rules! define_arena {
 
             #[inline(always)]
             /// Snapshot a position in the arena.
-            pub const fn mark(&self) -> $Mark { <$Mark>::new(self.len) }
+            $vis const fn mark(&self) -> $Mark { <$Mark>::new(self.len) }
             #[inline(always)]
             /// Rollback to a previous marked position.
-            pub const fn rollback(&mut self, m: $Mark) { self.len = m.0; }
+            $vis const fn rollback(&mut self, m: $Mark) { self.len = m.0; }
 
             /* byte slices */
 
             #[inline(always)]
             /// Returns a byte slice over all the written data.
-            pub const fn as_bytes(&self) -> &[u8] {
+            $vis const fn as_bytes(&self) -> &[u8] {
                 <$_>::slice_bytes(&self.data, 0, self.len)
             }
             /// Returns an exclusive byte slice over all the written data.
             #[inline(always)]
-            pub const fn as_bytes_mut(&mut self) -> &mut [u8] {
+            $vis const fn as_bytes_mut(&mut self) -> &mut [u8] {
                 $crate::__Arena::<CAP>::slice_bytes_mut(&mut self.data, 0, self.len)
             }
 
             /// Write a byte slice into the arena.
-            pub const fn push_bytes(&mut self, bytes: &[u8]) -> Option<$Handle> {
+            $vis const fn push_bytes(&mut self, bytes: &[u8]) -> Option<$Handle> {
                 $crate::unwrap!(some_if? self.len.checked_add(bytes.len()), |v| v <= CAP);
                 let start = self.len;
                 let handle = <$Handle>::new(start, bytes.len());
@@ -137,20 +140,20 @@ macro_rules! define_arena {
             }
 
             /// Read a shared slice over the written bytes.
-            pub const fn read_bytes(&self, h: $Handle) -> Option<&[u8]> {
+            $vis const fn read_bytes(&self, h: $Handle) -> Option<&[u8]> {
                 if h.offset() + h.len() > self.len { return None }
                 Some($crate::__Arena::<CAP>::slice_bytes(&self.data,
                     h.offset(), h.offset() + h.len()))
             }
             /// Read an exclusive slice over the written bytes.
-            pub const fn read_bytes_mut(&mut self, h: $Handle) -> Option<&mut [u8]> {
+            $vis const fn read_bytes_mut(&mut self, h: $Handle) -> Option<&mut [u8]> {
                 if h.offset() + h.len() > self.len { return None; }
                 Some($crate::__Arena::<CAP>::slice_bytes_mut(&mut self.data,
                     h.offset(), h.offset() + h.len()))
             }
 
             /// Replace the bytes for the handle. Lengths must match. Returns `false` otherwise.
-            pub const fn replace_bytes(&mut self, h: $Handle, new: &[u8]) -> bool {
+            $vis const fn replace_bytes(&mut self, h: $Handle, new: &[u8]) -> bool {
                 if let Some(dst) = self.read_bytes_mut(h) {
                     if dst.len() == new.len() {
                         dst.copy_from_slice(new);
@@ -163,7 +166,7 @@ macro_rules! define_arena {
             /* single bytes */
 
             /// Write a single byte into the arena.
-            pub const fn push_byte(&mut self, byte: u8) -> Option<$Handle> {
+            $vis const fn push_byte(&mut self, byte: u8) -> Option<$Handle> {
                 if self.len + 1 > CAP { return None; }
                 <$_>::write_byte(&mut self.data, self.len, byte);
                 let handle = <$Handle>::new(self.len, 1);
@@ -171,17 +174,17 @@ macro_rules! define_arena {
                 Some(handle)
             }
             /// Read a byte previously written.
-            pub const fn read_byte(&self, handle: $Handle) -> Option<u8> {
+            $vis const fn read_byte(&self, handle: $Handle) -> Option<u8> {
                 if handle.offset() + 1 > self.len { return None }
                 Some(<$_>::read_byte(&self.data, handle.offset()))
             }
             /// Read a byte previously written.
-            pub const fn read_byte_mut(&mut self, handle: $Handle) -> Option<&mut u8> {
+            $vis const fn read_byte_mut(&mut self, handle: $Handle) -> Option<&mut u8> {
                 if handle.offset() + 1 > self.len { return None }
                 Some(<$_>::read_byte_mut(&mut self.data, handle.offset()))
             }
             /// Replace the bytes for `handle`. Length must match.
-            pub const fn replace_byte(&mut self, handle: $Handle, new: u8) -> bool {
+            $vis const fn replace_byte(&mut self, handle: $Handle, new: u8) -> bool {
                 if handle.len() != 1 { return false; }
                 <$_>::write_byte(&mut self.data, handle.offset(), new);
                 true
@@ -193,7 +196,7 @@ macro_rules! define_arena {
             /// and spanning `count` items of its length.
             ///
             /// Returns `None` if...
-            pub const fn view_bytes(&self, h: $Handle, count: usize) -> Option<&[u8]> {
+            $vis const fn view_bytes(&self, h: $Handle, count: usize) -> Option<&[u8]> {
                 let total = h.len() * count;
                 if h.offset() + total > self.len { return None; }
                 Some(<$_>::slice_bytes(&self.data, h.offset(), h.offset() + total))
@@ -203,7 +206,7 @@ macro_rules! define_arena {
             /// and spanning `count` items of its length.
             ///
             /// Returns `None` if...
-            pub const fn view_bytes_mut(&mut self, h: $Handle, count: usize) -> Option<&mut [u8]> {
+            $vis const fn view_bytes_mut(&mut self, h: $Handle, count: usize) -> Option<&mut [u8]> {
                 let total = h.len() * count;
                 if h.offset() + total > self.len { return None; }
                 Some(<$_>::slice_bytes_mut(&mut self.data, h.offset(), h.offset() + total))
@@ -212,7 +215,7 @@ macro_rules! define_arena {
             /* shrinking the arena */
 
             /// Truncates the arena if the handle corresponds to the last region.
-            pub const fn truncate_last(&mut self, h: $Handle) -> bool {
+            $vis const fn truncate_last(&mut self, h: $Handle) -> bool {
                 let end = h.offset() + h.len();
                 if end != self.len { return false; }
                 self.len = h.offset();
@@ -220,7 +223,7 @@ macro_rules! define_arena {
             }
 
             /// Copies the last stored value into `dst` and removes it.
-            pub const fn pop_into(&mut self, h: $Handle, dst: &mut [u8]) -> bool {
+            $vis const fn pop_into(&mut self, h: $Handle, dst: &mut [u8]) -> bool {
                 if h.len() != dst.len() { return false; }
                 if let Some(src) = self.read_bytes(h) {
                     $crate::whilst! { i in 0..h.len(); { dst[i] = src[i]; }}
@@ -230,8 +233,8 @@ macro_rules! define_arena {
         }
     };
     // $_d: the dollar sign passed as a token, as a trick to be able to nest repetitions.
-    (%prim $Arena:ident, $Handle:ty, ($_d:tt)) => {
-        #[allow(dead_code)]
+    (%prim $vis:vis $Arena:ident, $Handle:ty, ($_d:tt)) => {
+        #[allow(dead_code, private_interfaces)]
         /// Implements push, read and replace for primitives.
         impl<const CAP: usize> $Arena<CAP> {
             impl_for_primitives!();
@@ -242,18 +245,18 @@ macro_rules! define_arena {
             ///
             /// # Errors
             /// Returns `None` if there’s insufficient capacity.
-            pub const fn push_bool(&mut self, val: bool) -> Option<$Handle> {
+            $vis const fn push_bool(&mut self, val: bool) -> Option<$Handle> {
                 self.push_byte(val as u8)
             }
             /// Reads a `bool` from the given `handle`.
             ///
             /// # Errors
             /// Returns `None` if the handle is invalid or incomplete.
-            pub const fn read_bool(&self, handle: $Handle) -> Option<bool> {
+            $vis const fn read_bool(&self, handle: $Handle) -> Option<bool> {
                 if let Some(b) = self.read_byte(handle) { Some(b != 0) } else { None }
             }
             /// Replaces a `bool` from the given `handle`. Returns `true` on success.
-            pub const fn replace_bool(&mut self, handle: $Handle, val: bool) -> bool {
+            $vis const fn replace_bool(&mut self, handle: $Handle, val: bool) -> bool {
                 self.replace_byte(handle, val as u8)
             }
 
@@ -264,7 +267,7 @@ macro_rules! define_arena {
             /// # Errors
             /// Returns `None` if there’s insufficient capacity.
             #[inline(always)]
-            pub const fn push_char(&mut self, val: char) -> Option<$Handle> {
+            $vis const fn push_char(&mut self, val: char) -> Option<$Handle> {
                 self.push_u32(val as u32)
             }
             /// Reads a `char` from the given `handle`.
@@ -272,12 +275,12 @@ macro_rules! define_arena {
             /// # Errors
             /// Returns `None` if the handle is invalid or incomplete.
             #[inline(always)]
-            pub const fn read_char(&self, handle: $Handle) -> Option<char> {
+            $vis const fn read_char(&self, handle: $Handle) -> Option<char> {
                 if let Some(c) = self.read_u32(handle) { char::from_u32(c) } else { None }
             }
             /// Replaces a `char` from the given `handle`. Returns `true` on success.
             #[inline(always)]
-            pub const fn replace_char(&mut self, handle: $Handle, val: char) -> bool {
+            $vis const fn replace_char(&mut self, handle: $Handle, val: char) -> bool {
                 self.replace_u32(handle, val as u32)
             }
         }
@@ -303,18 +306,18 @@ macro_rules! define_arena {
                 /// # Errors
                 /// Returns `None` if there’s insufficient capacity.
                 #[inline(always)]
-                pub const fn [<push_ $T>](&mut self, val: $T) -> Option<$Handle> {
+                $vis const fn [<push_ $T>](&mut self, val: $T) -> Option<$Handle> {
                     self.push_byte(val as u8)
                 }
                 #[doc = "Reads a `" $T "` from the given `handle`."]
                 ///
                 /// # Errors
                 /// Returns `None` if the handle is invalid or incomplete.
-                pub const fn [<read_ $T>](&self, handle: $Handle) -> Option<$T> {
+                $vis const fn [<read_ $T>](&self, handle: $Handle) -> Option<$T> {
                     if let Some(b) = self.read_byte(handle) { Some(b as $T) } else { None }
                 }
                 #[doc = "Replaces a `" $T "` from the given `handle`. Returns `true` on success."]
-                pub const fn [<replace_ $T>](&mut self, handle: $Handle, val: $T) -> bool {
+                $vis const fn [<replace_ $T>](&mut self, handle: $Handle, val: $T) -> bool {
                     self.replace_byte(handle, val as u8)
                 }
             }};
@@ -327,14 +330,14 @@ macro_rules! define_arena {
                 /// # Errors
                 /// Returns `None` if there’s insufficient capacity.
                 #[inline(always)]
-                pub const fn [<push_ $T>](&mut self, val: $T) -> Option<$Handle> {
+                $vis const fn [<push_ $T>](&mut self, val: $T) -> Option<$Handle> {
                     self.push_bytes(&val.to_le_bytes())
                 }
                 #[doc = "Reads a `" $T "` in little-endian order from the given `handle`."]
                 ///
                 /// # Errors
                 /// Returns `None` if the handle is invalid or incomplete.
-                pub const fn [<read_ $T>](&self, handle: $Handle) -> Option<$T> {
+                $vis const fn [<read_ $T>](&self, handle: $Handle) -> Option<$T> {
                     const T_SIZE: usize = size_of::<$T>();
                     if let Some(bytes) = self.read_bytes(handle) {
                         Some($T::from_le_bytes(
@@ -342,7 +345,7 @@ macro_rules! define_arena {
                     } else { None }
                 }
                 #[doc = "Replaces a `" $T "` from the given `handle`. Returns `true` on success."]
-                pub const fn [<replace_ $T>](&mut self, handle: $Handle, val: $T) -> bool {
+                $vis const fn [<replace_ $T>](&mut self, handle: $Handle, val: $T) -> bool {
                     if let Some(b) = self.read_bytes_mut(handle) {
                         if let Some(arr) = b.first_chunk_mut::<{size_of::<$T>()}>() {
                             *arr = val.to_le_bytes();
@@ -361,7 +364,7 @@ macro_rules! define_arena {
                 ///
                 /// # Errors
                 /// Returns `None` if there’s insufficient capacity or the string is too long.
-                pub const fn [<push_str_ $T>](&mut self, val: &str) -> Option<$Handle> {
+                $vis const fn [<push_str_ $T>](&mut self, val: &str) -> Option<$Handle> {
                     let len = val.len();
                     if len <= <$T>::MAX as usize {
                         let prefix = $crate::unwrap![some? self.[<push_ $T>](len as $T)];
@@ -375,7 +378,7 @@ macro_rules! define_arena {
                 ///
                 /// # Errors
                 /// Returns `None` if the handle is invalid or incomplete.
-                pub const fn [<read_str_ $T>](&self, handle: $Handle) -> Option<&str> {
+                $vis const fn [<read_str_ $T>](&self, handle: $Handle) -> Option<&str> {
                     let len_size = size_of::<$T>();
                     let handle = $Handle::new(handle.offset() + len_size, handle.len() - len_size);
                     let s = $crate::unwrap![some? self.read_bytes(handle)];
@@ -386,7 +389,7 @@ macro_rules! define_arena {
                 /// from the given `handle`. Returns `true` on success.
                 ///
                 /// Both strings have to have the same byte length.
-                pub const fn [<replace_str_ $T>](&mut self, handle: $Handle, val: &str) -> bool {
+                $vis const fn [<replace_str_ $T>](&mut self, handle: $Handle, val: &str) -> bool {
                     let len_size = size_of::<$T>();
                     let handle = $Handle::new(handle.offset() + len_size, handle.len() - len_size);
                     if handle.len() != val.len() { return false }