+++ libs/base_core/src/sys/mem/arena/mod.rs
@@ -4,12 +4,10 @@
 //
 
 mod define; // defina_arena!
-mod handle; // ArenaHandle
 mod internals;
 
 // WIPZONE
 // mod field; // ArenaField
-// mod handle_macro;
 // mod primitive; // ArenaPrimitive
 // mod recipe; // ArenaRecipe
 
@@ -17,10 +15,8 @@ crate::structural_mods! { // _mods, hidden
     _mods {
         pub use super::{
             define::*,
-            handle::*,
             internals::*,
             // WIPZONE
-            // handle_macro::*, // TODO
             // field::*,
             // primitive::*,
             // recipe::*,